backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.28
# Use SQLite for development
# psycopg2==2.9.9
alembic==1.13.1
# Use only OpenCV for face detection (simpler installation)
opencv-contrib-python==4.9.0.80
numpy==1.24.3
pillow==10.0.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0